adds noop header validator
