+++ method/client/verification/src/testFixtures/java/uk/co/idv/context/adapter/verification/client/header/NoopIdvHeaderValidator.java
@@ -1,0 +1,10 @@
+package uk.co.idv.context.adapter.verification.client.header;
+
+public class NoopIdvHeaderValidator implements IdvHeaderValidator {
+
+    @Override
+    public void validate(IdvRequestHeaders headers) {
+        // intentionally blank
+    }
+
+}
